Added i18n mechanism and locale file of Japanese.
You can change the value of "title", "text" or "name" property by chrome i18n message catalog.

ex)
title:
'About App Inspector for Sencha'
to
'App Inspector for Sencha について'

It allows the message key which contains space value. If doesn't exist a message in the catalog, use it just as it is.